fixed handler of errors, main function, schemas
