+++ controllers/contacts.controllers.js
@@ -5,7 +5,7 @@ const {
   removeContact,
   updateContactById,
 } = require("../services/contacts");
-const { WrongParametersError } = require("../helpers/index");
+const { HttpError } = require("../helpers/index");
 
 const getContacts = async (req, res) => {
   const contacts = await listContacts();
@@ -18,7 +18,7 @@ const getContact = async (req, res, next) => {
   const contact = await getContactById(contactId);
 
   if (!contact) {
-    throw new WrongParametersError(`Not found contact by id: ${contactId}`);
+    throw new HttpError(404, `Not found contact by id: ${contactId}`);
   }
   return res.json(contact);
 };
@@ -33,7 +33,7 @@ const deleteContact = async (req, res, next) => {
   const contact = await getContactById(contactId);
 
   if (!contact) {
-    throw new WrongParametersError(`Not found contact by id: ${contactId}`);
+    throw new HttpError(404, `Not found contact by id: ${contactId}`);
   }
   await removeContact(contactId);
   return res.status(200).json({ message: "contact deleted" });
@@ -44,7 +44,7 @@ const updateContact = async (req, res, next) => {
   const contact = await updateContactById(contactId, req.body);
 
   if (!contact) {
-    throw new WrongParametersError(`Not found contact by id: ${contactId}`);
+    throw new HttpError(404, `Not found contact by id: ${contactId}`);
   }
   return res.status(200).json(contact);
 };
@@ -54,7 +54,7 @@ const updateStatusContact = async (req, res, next) => {
   const contact = await updateContactById(contactId, req.body);
 
   if (!contact) {
-    throw new WrongParametersError(`Not found contact by id: ${contactId}`);
+    throw new HttpError(404, `Not found contact by id: ${contactId}`);
   }
   return res.status(200).json(contact);
 };
